fix: line ripple notched outline and floating label screenshot tests
